--- .vscode/settings.json
@@ -1,4 +1,0 @@
-{
-  "editor.tabSize": 2,
-  "editor.insertSpaces": true
-}
