+++ database/encrypt.py
@@ -1,12 +1,12 @@
 from argon2 import PasswordHasher
 
 def secure_password(password):
-  ph = PasswordHasher()
-  return ph.hash(password)
+	ph = PasswordHasher()
+	return ph.hash(password)
 
 def check_password(password, hashed_password):
-  ph = PasswordHasher()
-  try:
-    return ph.verify(hashed_password, password)
-  except:
-    return False
+	ph = PasswordHasher()
+	try:
+		return ph.verify(hashed_password, password)
+	except:
+		return False
