+++ database/localStorageServer.py
@@ -2,4 +2,4 @@ from streamlit_ws_localstorage import injectWebsocketCode
 import uuid
 
 def server():
-  return injectWebsocketCode(hostPort='wsauthserver.supergroup.ai', uid=str(uuid.uuid1()))
+	return injectWebsocketCode(hostPort='wsauthserver.supergroup.ai', uid=str(uuid.uuid1()))
