+++ src/apps/adminTools/packageUsed.py
@@ -13,5 +13,3 @@ def packageUsed():
   st.title("Package Used")
   packages = readPackages('requirements.txt')
   st.write(packages)
-
-packageUsed()
