+++ src/helpers/camelToReadable.py
@@ -1,3 +1,3 @@
 def camelToReadable(camelCaseName):
-  readableName = ''.join([' ' + char if char.isupper() else char for char in camelCaseName]).strip()
-  return readableName.title()
+	readableName = ''.join([' ' + char if char.isupper() else char for char in camelCaseName]).strip()
+	return readableName.title()
